Fix bug when sequence is defined but it is empty
The if statement with "is not none" only works if the sequence is not defined at all, so when it is defined and it is empty it will lead to compilation error due to printing "as edwSequence" without an input before. 
Using the macro datavault4dbt.is_something() fixes this issue.