Bugfix in PIT - Exasol version macro
In the PIT macro - Exasol version The dash in the closing curly brackets(line 70 and 76) led to a lack of whitespace between the first and second iteration of the loop(when there is more than one source model) in the line 71... so the model when having two sources would run into an error because the line 75 or 77 would be glued with the line 71 that starts the LEFT JOIN again. 
Removing these dashes of the closing curly brackets fixes the issue.